unionml_demo/requirements.txt
git+https://www.github.com/unionai-oss/unionml@main
flytekitplugins-onnxpytorch
gradio
plotly
tensorboardX
torch
tqdm
transformers
wandb
